fix the crash
we were often using 'tails' to produce one Stack for each element in the
list of entities, but this is not quite right since this includes one
final Stack which is an empty list, and a Stack is not supposed to be
empty. Almost everywhere, empty Stacks were simply filtered out, but in
this one place it was kept, leading to the 'head' error.

I should either implement a function which is almost like 'tails' but
doesn't include this empty Stack, or accept that Stacks can be empty and
be more careful with my calls to 'head'.