fix hang
for technical reasons, the 'pure' implementation for 'ZipList' produces
an infinite list. This is problematic when computing (or [False,False,...]).
This new API makes sure the list is finite before computing the 'or'.

but it's _still_ not right because entities are now moving _under_ other
entities instead of moving on top of them.